+++ SlowTest/someSlowTestParmsFile.h
@@ -73,7 +73,7 @@ public:
    int  mTestMode;
 
    // Initial value.
-   double  mInitialValue;
+   float  mInitialValue;
 
    //***************************************************************************
    //***************************************************************************
@@ -95,12 +95,12 @@ public:
    // Members. These are used int expand.
 
    // Fuzzy alpha filter sampling period and time constant, seconds.
-   double mAlphaFilterTs;
-   double mAlphaFilterTc;
+   float mAlphaFilterTs;
+   float mAlphaFilterTc;
 
    // Fuzzy to crisp threshold.
-   double mFuzzyToCrispThreshLo;
-   double mFuzzyToCrispThreshHi;
+   float mFuzzyToCrispThreshLo;
+   float mFuzzyToCrispThreshHi;
 
    //***************************************************************************
    //***************************************************************************
